*)again fixing for wrong version display, now totally working with double instead of float
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@2464 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ htroot/Status.java
@@ -137,7 +137,7 @@ public class Status {
         } else {
             prop.put("versioncomment", 0); // no comment
         }
-        prop.put("versioncomment_latestVersion", Float.toString(yacyCore.latestVersion));
+        prop.put("versioncomment_latestVersion", Double.toString(yacyCore.latestVersion));
 
         // hostname and port
         String extendedPortString = env.getConfig("port", "8080");
